% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{csafe_cols}
\alias{csafe_cols}
\title{Primary CSAFE colors}
\format{An object of class \code{character} of length 5.}
\usage{
csafe_cols
}
\description{
Primary CSAFE colors
}
\keyword{datasets}
