% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{csafe_cols_secondary}
\alias{csafe_cols_secondary}
\title{Secondary CSAFE colors}
\format{An object of class \code{character} of length 12.}
\usage{
csafe_cols_secondary
}
\description{
Secondary CSAFE colors
}
\keyword{datasets}
