% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isu-colors.R
\docType{data}
\name{isu_cols_secondary}
\alias{isu_cols_secondary}
\title{Secondary ISU colors}
\format{An object of class \code{character} of length 15.}
\usage{
isu_cols_secondary
}
\description{
Secondary ISU colors
}
\keyword{datasets}
