% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-color.R
\name{scale_color_csafe}
\alias{scale_color_csafe}
\title{CSAFE scale color for ggplot2}
\usage{
scale_color_csafe(theme = "primary", csafe_key = list(primary = c("#40B4E5",
  "#003A70", "#77BC1F", "#CF0A2C", "#8A8A8D"), secondary = c("#9F1B96",
  "#EA1D76", "#AC162C", "#00833E", "#E11383", "#E4E833", "#F5BD47", "#555759",
  "#2E5597", "#6BA2B9", "#FA8D29", "#005487")))
}
\description{
csafe scale color
}
\examples{
scale_color_csafe(theme="primary")
}
\author{
Sam Tyner
}
