% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-color.R
\name{scale_color_isu}
\alias{scale_color_isu}
\title{ISU scale color for ggplot2}
\usage{
scale_color_isu(theme = "primary", isu_key = list(primary = c("#C8102E",
  "#F1BE48", "#524727", "#9B945F", "#CAC7A7"), secondary = c("#3E4827",
  "#76881D", "#A2A569", "#003D4C", "#006BA6", "#7A99AC", "#7C2529", "#9A3324",
  "#BE531C", "#8B5B29", "#B9975B", "#EED484", "#6E6259", "#707372", "#ACA39A")))
}
\description{
isu scale color
}
\examples{
scale_color_ISU(theme="primary")
}
\author{
Sam Tyner
}
