% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-color.R
\name{scale_color_malisha}
\alias{scale_color_malisha}
\title{Malisha's scale color for ggplot2}
\usage{
scale_color_malisha(...)
}
\description{
csafe scale color
}
\examples{
scale_color_malisha()
}
\author{
Sam Tyner
}
