% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_csafe}
\alias{theme_csafe}
\title{CSAFE themes for ggplot2}
\usage{
theme_csafe(theme = "primary", csafe_key = list(primary = list(family_title
  = "Montserrat Bold", family_text = "Montserrat Light", colour_title =
  "#003A70", colour_text = "#8A8A8D"), secondary = list(family_title =
  "Montserrat Bold", family_text = "Montserrat Light", colour_title = "#9F1B96",
  colour_text = "#555759")))
}
\description{
csafe theme
}
\examples{
theme_csafe(theme="primary")
}
\author{
Sam Tyner
}
