% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_isu}
\alias{theme_isu}
\title{ISU themes for ggplot2}
\usage{
theme_isu(theme = "primary", isu_key = list(primary = list(family_title =
  "Univers Bold", family_text = "Univers Light", colour_title = "#524727",
  colour_text = "#9B945F"), secondary = list(family_title = "Univers Bold",
  family_text = "Univers Light", colour_title = "#7C2529", colour_text =
  "#6E6259")))
}
\description{
isu theme
}
\examples{
ggplot(data = mtcars) +
    geom_point(aes(x = drat, y = qsec)) + 
    theme_isu(theme="primary")
}
\author{
Sam Tyner
}
